Expanded new encoding system to be more generic and easier to write definitions for

+++ src/common/com_debug.cpp
@@ -15,7 +15,7 @@ inline void COM_PrintBits(i32 val, u8 printNewLine)
     }
 }
 
-inline void COM_PrintBytes(u8* bytes, u16 numBytes, i32 bytesPerRow)
+inline void COM_PrintBytes(u8* bytes, i32 numBytes, i32 bytesPerRow)
 {
     u8* read = bytes;
     if (bytesPerRow <= 0)
